refactor: 장바구니에서 상품 삭제 시 UX 개선
alert 창을 띄워 사용자의 의사를 한번더 확인

+++ src/hooks/useCart.ts
@@ -38,11 +38,14 @@ export const useCart = () => {
   };
 
   const handleRemoveCheckedItem = () => {
-    checkedItems.forEach((id) => {
-      api.delete(`${CART_BASE_URL}/${id}`, { id });
-      removeProductItemFromCart(id);
-    });
-    setCheckedItems([]);
+    const confirmResult = window.confirm('정말로 삭제하시겠습니까?');
+    if (confirmResult) {
+      checkedItems.forEach((id) => {
+        api.delete(`${CART_BASE_URL}/${id}`, { id });
+        removeProductItemFromCart(id);
+      });
+      setCheckedItems([]);
+    }
   };
 
   return {
